Implement programmatic page splitting
This approach doesn't work as well as I expected it to.

The final text that gets parsed out is super segmented and not really sensible. Not to mention it treats the codes (S1, S2, S2.1, S2.2, etc) as a separate column and the actual information as another column, this is not what we want.